requirements.txt
streamlit
Pillow
nltk
torch
python-dotenv
google-generativeai
ultralytics     
